use ANA_MSG_HEADER/SOURCE for standalone code with no msg()
