+++ Root/TracksInJetHists.cxx
@@ -1,10 +1,10 @@
 #include "xAODAnaHelpers/TracksInJetHists.h"
 #include <math.h>
 #include <xAODTracking/TrackParticle.h>
-#include <AsgTools/MessageCheck.h>
 #include "xAODAnaHelpers/HelperFunctions.h"
 
 using std::cout;  using std::endl;
+ANA_MSG_SOURCE(msgTracksInJetHists, "TracksInJetHists")
 
 TracksInJetHists :: TracksInJetHists (std::string name, std::string detailStr) :
   HistogramManager(name, detailStr)
@@ -62,7 +62,7 @@ void TracksInJetHists::record(EL::Worker* wk) {
 
 
 StatusCode TracksInJetHists::execute( const xAOD::TrackParticle* trk, const xAOD::Jet* jet,  const xAOD::Vertex *pvx, float eventWeight,  const xAOD::EventInfo* eventInfo ) {
-
+  using namespace msgTracksInJetHists;
   //
   //  Fill track hists
   //
